feat: update file and folder structure and rename `UserProfile` to `Complete UserProfile`
