+++ src/frontend/src/components/CompleteUserProfile/FormContents/OtherDetails/index.tsx
