+++ src/frontend/src/components/CompleteUserProfile/UserProfileHeader/index.tsx
